requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
aiohttp=3.7.4.post0=pypi_0
ase=3.20.1=pypi_0
async-timeout=3.0.1=pypi_0
attrs=20.3.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.9.30=h06a4308_1
cached-property=1.5.2=pypi_0
cachetools=4.2.2=pypi_0
cairo=1.14.12=h8948797_3
certifi=2021.5.30=py36h06a4308_0
chardet=4.0.0=pypi_0
clearml=1.0.2=pypi_0
clearml-agent=0.17.2=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.24=pypi_0
dataclasses=0.8=pypi_0
decorator=4.4.2=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.9.0=pypi_0
furl=2.1.2=pypi_0
future=0.18.2=pypi_0
glib=2.66.1=h92f7085_0
google-auth=1.30.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
googledrivedownloader=0.4=pypi_0
grpcio=1.38.0=pypi_0
h5py=3.1.0=pypi_0
humanfriendly=9.1=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
idna-ssl=1.1.0=pypi_0
importlib-metadata=4.0.1=pypi_0
intel-openmp=2020.2=254
ipykernel=5.5.5=pypi_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
isodate=0.6.0=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=1.0.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-core=4.7.1=pypi_0
kiwisolver=1.3.1=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libboost=1.73.0=h37e3b65_11
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
littleutils=0.2.2=pypi_0
llvmlite=0.35.0=pypi_0
lz4-c=1.9.2=heb0550a_3
markdown=3.3.4=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.3=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
monty=2021.6.10=pypi_0
mpmath=1.2.1=pypi_0
multidict=5.1.0=pypi_0
ncurses=6.2=he6710b0_1
networkx=2.5=pypi_0
ninja=1.10.2=py36hff7bd54_0
numba=0.52.0=pypi_0
numpy=1.19.4=pypi_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.1=pypi_0
ogb=1.3.1=pypi_0
olefile=0.46=py_0
openssl=1.1.1l=h7f8727e_0
orderedmultidict=1.0.1=pypi_0
outdated=0.2.1=pypi_0
packaging=21.0=pypi_0
palettable=3.3.0=pypi_0
pandas=1.1.5=pypi_0
parso=0.7.1=pypi_0
pathlib2=2.3.5=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.0.1=py36he98fc37_0
pip=20.3.3=py36h06a4308_0
pixman=0.40.0=h7b6447c_0
prompt-toolkit=3.0.18=pypi_0
protobuf=3.17.2=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pypi_0
py-boost=1.73.0=py36h962f231_11
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeprecate=0.3.0=pypi_0
pydispatcher=2.0.5=pypi_0
pygments=2.9.0=pypi_0
pyhocon=0.3.57=pypi_0
pyjwt=1.7.1=pypi_0
pymatgen=2020.1.10=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python-louvain=0.15=pypi_0
pytorch=1.7.1=py3.6_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=1.3.0=pypi_0
pytorch-metric-learning=0.9.99=pyh39e3cac_0
pytorch-warmup=0.0.4=pypi_0
pytz=2020.5=pyhd3eb1b0_0
pyyaml=5.3.1=pypi_0
pyzmq=22.1.0=pypi_0
rdflib=5.0.0=pypi_0
rdkit=2020.09.1.0=py36hd50e099_1
rdkit-pypi=2021.3.5.1=pypi_0
readline=8.0=h7b6447c_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
ruamel-yaml=0.17.9=pypi_0
ruamel-yaml-clib=0.2.2=pypi_0
scikit-learn=0.24.0=pypi_0
scipy=1.5.4=pypi_0
setuptools=51.0.0=py36h06a4308_2
six=1.15.0=py36h06a4308_0
spglib=1.16.1=pypi_0
sqlite=3.33.0=h62c20be_0
sympy=1.8=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.2=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
torch-cluster=1.5.8=pypi_0
torch-geometric=1.7.0=pypi_0
torch-scatter=2.0.5=pypi_0
torch-sparse=0.6.8=pypi_0
torch-spline-conv=1.2.0=pypi_0
torchaudio=0.7.2=py36
torchmetrics=0.5.1=pypi_0
torchvision=0.8.2=py36_cu101
tornado=6.1=pypi_0
tqdm=4.55.0=pypi_0
traitlets=4.3.3=pypi_0
typing=3.7.4.3=pypi_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pypi_0
virtualenv=16.7.10=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=pypi_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
